Correctly request & process JSON progress output, "tag" other stdout
Switched from select to the higher-level (and more performant) selectors.
Apparently I wan't getting the progress updates before because --log-json is
a different option than --json (:-P). It also uses a PTY in nonblocking mode
to detect and show prompts, along with saving the answers so multiple backup
processes are automatically answered the same way the user did. Also, other
output from Borg is prepended with which instance caused it (which is extra
useful if --stats is enabled).