Add dummy implementation of DiskLock
I need to do more research on how virtlockd works to get this to work.
It looks like the best way to implement it would be by requesting a lock
from the running daemon, but flocking the files by itself is definitely
a simpler strategy and it might work fine for local disks (see issue #7).