Add restore-vm (#1)
The parser for restore-vm is a bit simpler because it only has positional
arguments and flags at the beginning like a normal program (no --borg-args
shenanigans). It can actually reuse the ArchiveBuilder class for the restore
because the bind mounts are not read-only, but this doesn't work if the
format of the disk changes.